Refactor: 인증 서비스 마이그레이션
